double to decimals in AccountBalanceCalc

+++ c-sharp-instruction/joe_hawes_account_balance_calculator/joe_hawes_account_balance_calculator/Model/CheckingAccount.cs
@@ -4,8 +4,8 @@ namespace joe_hawes_account_balance_calculator.Model
 {
 	public class CheckingAccount : Account
 	{
-		public double MonthlyFee { get; set; }
-		public CheckingAccount(double initialBalance, double fee) : base(initialBalance)
+		public decimal MonthlyFee { get; set; }
+		public CheckingAccount(decimal initialBalance, decimal fee) : base(initialBalance)
 		{
 			MonthlyFee = fee;
 		}
