+++ c-sharp-instruction/joe_hawes_account_balance_calculator/joe_hawes_account_balance_calculator/Model/SavingsAccount.cs
@@ -4,10 +4,10 @@ namespace joe_hawes_account_balance_calculator.Model
 {
 	public class SavingsAccount : Account
 	{
-		public double MonthlyInterestRate { get; set; }
-		public double MonthlyInterestPayment { get; private set; }
+		public decimal MonthlyInterestRate { get; set; }
+		public decimal MonthlyInterestPayment { get; private set; }
 
-		public SavingsAccount(double initialBalance, double interestRate) :  base(initialBalance)
+		public SavingsAccount(decimal initialBalance, decimal interestRate) :  base(initialBalance)
 		{ 
 			MonthlyInterestRate = interestRate;
